Add GitHub workflow for creating builder
A new workflow, "process-builder.yaml", has been added in .github/workflows, which is triggered on push events on main and develop branches except when changes are made in the "containerfiles" directory. This workflow runs on ubuntu-latest and includes steps to checkout code and build and publish using "pack builder create".