Update create-builder workflow to download pack CLI
The workflow has been modified specifically for the GitHub Actions. The 'pack' CLI is now downloaded into a new 'bin' directory. The PATH is updated to include the new directory improving the organization and management of downloaded files. Moreover, experimental features have been enabled in the pack configuration for enhanced functionality.